v2/app/requirements.txt
npyscreen==4.10.5
numpy
opencv-python
python-dotenv
python-telegram-bot